Feature sine wave display
Biorhythm theory is based on three sine waves of different frequencies originating at birth.

Add a feature to display these sine waves on a graph with the X axis for time, covering a week before and a week after the selected date.

This will (a) illustrate the effect of sine waves of different frequencies, and (b) provide more visual data for building a confirmation bias.

Adds optional non-Sugar platform dependencies.  Please install the libraries python-scipy, python-numpy and python-matplotlib through apt or your favourite package manager, to display the line graph.  When dependencies are missing, the activity operates properly but does not display the line graph.

Closes https://github.com/sugarlabs/biorhythm/issues/9